Remove handPickedCollections code out of Home page into seperate component
